Add signal handler and resume option
This is based on the examples from

https://gobyexample.com/signals
https://gobyexample.com/select
https://gobyexample.com/closing-channels

. It's been a while since I worked with Go channels, but the overall
process is straight-forward.

We can use `signal.Notify` to listen to certain OS signals and send
these signals to a channel to be processed. For now we check for
`SIGINT` and `SIGTERM`, both of which indicate that the process has been
intentionally signalled to stop. I think we can ignore other signals as
the most common way to stop the program will be someone pressing CTRL +
c.

We use another channel to send back the ID of the subtitle we're
currently processing. The information is stored in a variable and
updated with each subtitle we sync. When we detect a `SIGINT` or
`SIGTERM` signal, we print a message with the command to continue with a
Radarr or Sonarr ID and exit the program.

Using this ID is a compromise. It's much more robust than using the name
of the subtitle file, but also ignores the fact that a movie or an
episode of a show can have multiple subtitle files with different
languages, some of which might have already been processed. I think it's
fair to ignore this for now, because syncing the subtitles again will
just take a few seconds. This would only pose a real problem if we're
dealing with a huge number of subtitle files, but even so we can save a
significant amount of time by skipping over already processed movies or
episodes.

Go's `select` statement gives us an elegant way to interact with both
channels mentioned above and even allows us to check if a channel has
been closed. We use the latter to check if all subtitles have been
processed and if so, do a clean exit.

For all of this to work, we need to run the actual subtitle sync as a
background process. Luckily this is incredibly easy using Go's built in
Goroutines. Just putting the `go` statement in front of a function is
enough to have it run concurrently in the background.